fix: make fillMissingData request specific routes when asking npm metadata
There are certain npm packages, (e.g:
https://registry.npmjs.org/@typescript-eslint/eslint-plugin)
that are extremely heavy in terms of number of versions and metadata
in general.

NPM.fillMissingData was asking for the full list when called causing
issues to the caller when the timeout is reached.

This fix makes fillMissingData call the specific routes for what it
needs so we remove the number of versions from the equation.